Source Update 15.10.10.0 - DEBUG: Added MERLIN mode for debugging, though the mode itself does not yet work, it can already be toggled though. - DEBUG: And the same goes for the RAMBO mode. - Close #67 - This works perfectly now. - Close #218 - This was one hell of a job as stuff did not easily resolve. Not only that, but I was also dependent on how the game would TIME things on me, and I wasnt very lucky with that one. But it is done now, and I am happy about that. - FIXED: A bug not listed in the issue tracker was a texture reference error in the corridor of Excalibur when you meet the three Cyborgs there. Due to the enemy picture move this bug was caused, and as I was already past that point content wise this bug could slumber in, and it was shown because it was the perfect place to test #218. It is fixed now, so the crash happening there now belongs to the past. - FIXED: A bug in the enemy AI did make it impossible to fix #218, so that one is fixed now too. - FIXED: The Cyborg Medic was only supposed to use HEALING items and yet she was throwing all sorts of attacks at me. This is a result of bug #203 and even after that one is fixed some results may come my way. The Cyborg Medic is fixed though.
Well, and that will be one of my biggest commits and also my final today. It is almost 2am and was not planning to make it this late. I am going to take a shower and relax. My next focus will be the development of the inside of the bar and that will be one hell of a job due to the graphic style of the bar itself, the need of the customers and of course, the introduction of the playable character Yirl. Expect me to need the entire week at least to get this all done.