Alpha Update # 87 Sun 18 October 2015; 01:22:29 CET Most important stuff in the new build: - Fixing a multi-map save bug - Artwork for Foxy - A few preparations to make the new game+ possible - #111 is now properly covered. Overgrinding will get you nowhere. - #113 is fixed - Foxy was shooting. That was a bug, as she's intended to use daggers. This is fixed. - It is now possible to advance to room 003 in the dungeon, however, you cannot do anything here at the present time. This location will be the host of the first puzzle in the game.
And that covers the most important shit in this build. We are well on the way. To let you into a secret, this dungeon is the last dungeon to pass through prior to the final dungeon of the demo. (An optional dungeon I have still planned not counted). So the idea of having the demo finished before Christmas does become quite realistic, however, no promises yet. I still have a lot of issues on the issue tracker to work down until the demo can be released and some of them can be really time eating. And especially the - kick Reggie - animation is something I am really worried about. I already have an approach in mind, but it is kinda tricky.

So far my story for today.