Revealed music artist list
This is the first version of the list. After this commit this list will be automatically updated each time a piece of music is added, modified or removed. This list is thus as such never a final list until the final release of te game.
The game itself will contain a 'jukebox' feature which you can unluck after completing a certain (optional) dungeon, and in there you can also listen to all pieces of music and see their respective artists.
License information about all tracks can be found if you analyse the 'Star Story.JCR' file with programes like GJCR.