Source Update 16.07.08.0 Right,
I've now setup the 'top line' of the final dungeon.
Meaning that it shows the name of the current area, as well as the keycards you have found in this area.
Keycards are area bound, so you cannot take them from one area to another, but you will get them back when you get back to the area. That is the basic idea.

All this works completes and closes #486